pyproject.toml
[project]
name = "traces"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.86.0",
    "pillow>=12.1.1",
]
